Reorganize flat modules into data/ indicators/ mailer/ packages
- data/: get_coin_data, get_coin_names, get_market_cap, get_fees, get_portfolio, google_trends
- indicators/: rsi, rolling_mean, bollinger_bands
- mailer/: mails, jinja_customs, mail_templates/
- Update all cross-module imports + mail template paths accordingly
- README: document package layout and run-from-root commands
- root keeps entry points (main.py, backtest.py) + constants.py, config.py

Verified: compileall passes, no stale flat imports, all package paths resolve.
