+++ mailer/mail_templates/coin_template.html
